requirement.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
generativeai==0.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
openai==1.64.0
packaging==24.2
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
python-telegram-bot==21.10
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3